+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - [breaking-change] `Defaults` field renamed on `RegisterProperties`
   and added into `Peripheral` and `ClusterInfo`
+- [breaking-change] `Field` splitted on `Field` enum and `FieldInfo` struct
+  to support field arrays
 - Added `derived_from` into `Field` and `ClusterInfo`
 - Updated dependencies, use `Edition 2018`
 - Added missing `zeroToToggle`
